+++ srcs/ft_print_upper_x.c
@@ -14,10 +14,10 @@
 
 int	ft_print_upper_x(va_list args)
 {
-	uint64_t	nbr;
-	uint64_t	size;
+	size_t	nbr;
+	size_t	size;
 
-	nbr = va_arg(args, uint64_t);
+	nbr = va_arg(args, size_t);
 	ft_putnbr_base_unsigned((unsigned int)nbr, "0123456789ABCDEF");
 	size = ft_get_size_num_u((unsigned int)nbr, 16);
 	return (size);
